move extract method to user service
