+++ backend/LexBoxApi/Services/UserService.cs
@@ -1,5 +1,7 @@
-﻿using LexBoxApi.Auth;
+﻿using System.Net.Mail;
+using LexBoxApi.Auth;
 using LexBoxApi.Services.Email;
+using LexCore.Exceptions;
 using LexData;
 using Microsoft.EntityFrameworkCore;
 
@@ -49,5 +51,36 @@ public class UserService(LexBoxDbContext dbContext, IEmailService emailService, 
         return (await dbContext.Users.Where(u => u.Id == id)
             .Select(u => u.UpdatedDate)
             .SingleOrDefaultAsync()).ToUnixTimeSeconds();
+    }
+
+    public static (string name, string? email, string? username) ExtractNameAndAddressFromUsernameOrEmail(string usernameOrEmail)
+    {
+        var isEmailAddress = usernameOrEmail.Contains('@');
+        string name;
+        string? email;
+        string? username;
+        if (isEmailAddress)
+        {
+            try
+            {
+                var parsed = new MailAddress(usernameOrEmail);
+                email = parsed.Address;
+                username = null;
+                name = parsed.DisplayName;
+                if (string.IsNullOrEmpty(name)) name = email.Split('@')[0];
+            }
+            catch (FormatException)
+            {
+                // FormatException message from .NET talks about mail headers, which is confusing here
+                throw new InvalidEmailException("Invalid email address", usernameOrEmail);
+            }
+        }
+        else
+        {
+            username = usernameOrEmail;
+            email = null;
+            name = username;
+        }
+        return (name, email, username);
     }
 }
